Move type inference code to its own module
